refactor: Meter 및 Distance에 Float를 인자로 받는 부생성자 추가

+++ android/app/src/main/java/io/coursepick/coursepick/data/location/DefaultLocationRepository.kt
@@ -77,7 +77,7 @@ class DefaultLocationRepository(
                     if (isFineLocationPermissionGranted) {
                         Location.Fine(coordinate)
                     } else {
-                        Location.Coarse(coordinate, Distance(location.accuracy.toDouble()))
+                        Location.Coarse(coordinate, Distance(location.accuracy))
                     },
                 )
             }.addOnFailureListener { exception: Exception ->
@@ -132,7 +132,7 @@ class DefaultLocationRepository(
                         } else {
                             Location.Coarse(
                                 coordinate,
-                                Distance(location.accuracy.toDouble()),
+                                Distance(location.accuracy),
                             )
                         },
                     )
